Wasm: share a single linear memory for the C runtimes
The non-wasi runtime declared three linear memories: jsstring.wat's
caml_buffer scratch buffer, plus one each from the separately-linked
zstd-impl.wasm and blake2-impl.wasm. Safari does not support modules
with several memories.

Link the zstd decoder and the blake2b hash into a single module
(runtime/wasm/c/, built as c-impl.wasm) that owns one memory, laid out
above the first 64 KiB page (--no-stack-first --global-base=65536) so
jsstring.wat can keep using that page from offset 0 and re-export it as
caml_buffer. zstd.wat and blake2.wat import the shared memory too. The
non-wasi runtime is now down to a single memory; the wasi runtime keeps
libc's memory separate (two memories).

The two C wrappers dropped their duplicated memcpy/memset/... stubs into
a shared stubs.c, and their build rules are merged into one declared via
a (subdir c ...) stanza.
