+++ runtime/wasm/c/blake2.h
